fixed re-renders
